Assignment 5
Your program should take in both user inputs using two scanf() statements as shown in the
sample. The inputs cannot be empty or newlines. Our assumption is that they will be positive
integers.
When the input values are changed to 7 and 19, the following output is produced:
FAST Main Building in ASCII Art
Enter number of floors: 7
Enter width of building: 19
!
(***)
|O|O|O|

/\______________________---------______________________/\
| || O O O O O O O || O || O O O || O || O O O O O O O || |
| || O O O O O O O || O || O O O || O || O O O O O O O || |
| || O O O O O O O || O || O O O || O || O O O O O O O || |
| || O O O O O O O || O || O O O || O || O O O O O O O || |
========================== _____ ==========================
| || O O O O O O O || O || | | || O || O O O O O O O || |
| || O O O O O O O || O || | | || O || O O O O O O O || |
| || O O O O O O O || O || |___| || O || O O O O O O O || |
/ / ___ \ \
/ * * * * * * * * * * * * / /___\ \ * * * * * * * * * * * * \
The width of the building will be an odd number. If the user enters an even number, the closest
odd number smaller than the user input will be chosen as the width. So, 7, 19 and 7, 20 will
produce the same output given above. The minimum width of the building is 3. If the user
enters a width less than 3, the base structure is printed out as follows.
FAST Main Building in ASCII Art
Enter number of floors: 4
Enter width of building: 2
_____
| |
| |
____ |___| ____
/ / ___ \ \
/ * / /___\ \ * \

This assignment is meant to give you practice with problem decomposition, loop statements,
nested loops and conditionals. You may use all the statements and constructs that you have
learnt unilt now.
The number of floors in the building design can also vary. The designers at FAST considered
various heights of the building. Initially a 2 storey building was constructed without a dome
as can be seen in an earlier image of FAST. So any construction of fewer than 3 floors would
not have a dome on it. It was also decided that if the building rises to eight or more floors it

would feature two water tanks on the top corners of the building (just like we have in the new
building!). We will stick to the same plan!
This means that if the user inputs eight or more floors, two tanks will be printed instead of the
dome structure as seen below.
FAST Main Building in ASCII Art
Enter number of floors: 8
Enter width of building: 10
______ ______
| | | |
| |___________________________| |
| || O O || O || O O O || O || O O || |
| || O O || O || O O O || O || O O || |
| || O O || O || O O O || O || O O || |
| || O O || O || O O O || O || O O || |
| || O O || O || O O O || O || O O || |
================ _____ ================
| || O O || O || | | || O || O O || |
| || O O || O || | | || O || O O || |
| || O O || O || |___| || O || O O || |
/ / ___ \ \
/ * * * * * * * / /___\ \ * * * * * * * \

However, if the height of the building is less than three floors, it will not feature any of these
structures as shown below.

FAST Main Building in ASCII Art
Enter number of floors: 2
Enter width of building: 17
_____

________________________ | | ________________________
| || O O O O O O || O || | | || O || O O O O O O || |